Change learning_rate param to float

+++ src/model.py
@@ -211,7 +211,7 @@ class TwoTower(pl.LightningModule):
     @staticmethod
     def add_model_specific_args(parent_parser):
         parser = ArgumentParser(parents=[parent_parser], add_help=False)
-        parser.add_argument('--learning_rate', type=int, default=1e-3)
+        parser.add_argument('--learning_rate', type=float, default=1e-3)
         parser.add_argument('--batch_size', type=int, default=32)
         parser.add_argument('--dim', type=int, default=128)
         parser.add_argument('--epochs', type=int, default=None)
